refactor : loginContent,registerContent를 포함하는 login 컴포넌트 이름을 auth로 바꿈
