+++ frontend/src/views/components/auth/registerContent.js
